Phase out the last remnant of JSON serialization by merging JsonSerializerShim into CsonSerializer.
